refactor pie chart for clarity
