+++ src/App.tsx
@@ -4,18 +4,14 @@ import { QueryStringHandler } from './storage/QueryStringHandler'
 import EditableField, { type EditableFieldHandle } from './EditableField'
 import './App.css'
 import Modal from './Modal'
+import { PieChart } from './PieChart'
 import {
-  polarToCartesian,
-  describeRingArc,
-  calculateAngles,
-  calculateRotation,
-} from './utils/pie'
-import {
-  useAnimatedNumber,
-  useAnimatedRadii,
-  useRevealOnChange,
-  usePrevious,
-} from './utils/animation'
+  getTaskByPath,
+  updateTaskAtPath,
+  addTaskAtPath,
+  deleteTaskAtPath,
+} from './utils/tasks'
+import type { Task } from './types'
 
 const VERSION =
   document.querySelector<HTMLMetaElement>('meta[name="build-version"]')?.content ||
@@ -24,262 +20,6 @@ const VERSION =
 const driveService = new GoogleDriveService()
 const queryHandler = new QueryStringHandler()
 
-interface Task {
-  id: number
-  name: string
-  description: string
-  effort: number
-  completed: boolean
-  subtasks: Task[]
-}
-
-function useIsMobile() {
-  const query = '(max-width: 600px)'
-  const [isMobile, setIsMobile] = useState(() =>
-    typeof window === 'undefined' ? false : window.matchMedia(query).matches,
-  )
-
-  useEffect(() => {
-    const mq = window.matchMedia(query)
-    const handler = (e: MediaQueryListEvent) => setIsMobile(e.matches)
-    mq.addEventListener('change', handler)
-    return () => mq.removeEventListener('change', handler)
-  }, [])
-
-  return isMobile
-}
-
-function getTasksAtPath(tasks: Task[], path: number[]): Task[] {
-  let list = tasks
-  for (const id of path) {
-    const t = list.find((x) => x.id === id)
-    if (!t) return []
-    list = t.subtasks
-  }
-  return list
-}
-
-function getTaskByPath(tasks: Task[], path: number[]): Task | null {
-  let list = tasks
-  let current: Task | null = null
-  for (const id of path) {
-    current = list.find((x) => x.id === id) || null
-    if (!current) return null
-    list = current.subtasks
-  }
-  return current
-}
-
-function updateTaskAtPath(
-  tasks: Task[],
-  path: number[],
-  updater: (t: Task) => Task,
-): Task[] {
-  if (path.length === 0) return tasks
-  const [id, ...rest] = path
-  return tasks.map((t) => {
-    if (t.id !== id) return t
-    if (rest.length === 0) return updater(t)
-    return { ...t, subtasks: updateTaskAtPath(t.subtasks, rest, updater) }
-  })
-}
-
-function addTaskAtPath(tasks: Task[], path: number[], task: Task): Task[] {
-  if (path.length === 0) return [...tasks, task]
-  const [id, ...rest] = path
-  return tasks.map((t) => {
-    if (t.id !== id) return t
-    return { ...t, subtasks: addTaskAtPath(t.subtasks, rest, task) }
-  })
-}
-
-function deleteTaskAtPath(tasks: Task[], path: number[]): Task[] {
-  if (path.length === 1) return tasks.filter((t) => t.id !== path[0])
-  const [id, ...rest] = path
-  return tasks.map((t) => {
-    if (t.id !== id) return t
-    return { ...t, subtasks: deleteTaskAtPath(t.subtasks, rest) }
-  })
-}
-
-function PieChart({
-  tasks,
-  path,
-  onSelect,
-  onUp,
-}: {
-  tasks: Task[]
-  path: number[]
-  onSelect: (p: number[]) => void
-  onUp: () => void
-}) {
-  const isMobile = useIsMobile()
-  const prevPath = usePrevious(path) || path
-  const depthDiff = path.length - prevPath.length
-  const parentFadeRaw = useRevealOnChange(path.length, 1500)
-
-  const prevParentPath = prevPath.slice(0, -1)
-  const prevCurrentTasks = getTasksAtPath(tasks, prevParentPath)
-  const prevAngles = calculateAngles(prevCurrentTasks)
-  const prevSelectedId = prevPath[prevPath.length - 1] ?? null
-  const prevSelectedIndex = prevCurrentTasks.findIndex(
-    (t) => t.id === prevSelectedId,
-  )
-  const prevRotationDeg =
-    (prevSelectedIndex >= 0
-      ? (calculateRotation(prevAngles[prevSelectedIndex].mid, isMobile) * 180) /
-        Math.PI
-      : 0)
-
-  const parentPath = path.slice(0, -1)
-  const currentTasks = getTasksAtPath(tasks, parentPath)
-  const selectedId = path[path.length - 1] ?? null
-  const angles = calculateAngles(currentTasks)
-  const selectedIndex = currentTasks.findIndex((t) => t.id === selectedId)
-  const targetRotation =
-    selectedIndex >= 0 ? calculateRotation(angles[selectedIndex].mid, isMobile) : 0
-  const rotation = useAnimatedNumber(targetRotation, 1500)
-  const rotationDeg = (rotation * 180) / Math.PI
-  const selectedTask = selectedIndex >= 0 ? currentTasks[selectedIndex] : null
-  const childTasks = selectedTask ? selectedTask.subtasks : []
-  const childAngles = calculateAngles(childTasks)
-  
-  const r1 = 70
-  const r2 = 100
-  const parentRadius = r1 * 0.4
-
-  const targetParent = path.length > 0 ? parentRadius : 0
-  const targetCurrent = { inner: path.length > 0 ? parentRadius : 0, outer: r1 }
-  const targetChild = { inner: r1 + 5, outer: r2 }
-
-  let fromParent: number | undefined
-  let fromCurrent: { inner: number; outer: number } | undefined
-  let fromChild: { inner: number; outer: number } | undefined
-  let fromPrevRing: { inner: number; outer: number } | undefined
-  let targetPrevRing: { inner: number; outer: number } = { inner: 0, outer: 0 }
-
-  if (depthDiff > 0) {
-    // drilling in
-    fromParent = r1
-    fromCurrent = { inner: r1 + 5, outer: r2 }
-    fromChild = { inner: r2 + 5, outer: r2 + 10 }
-    fromPrevRing = { inner: parentRadius, outer: r1 }
-    targetPrevRing = { inner: 0, outer: parentRadius }
-  } else if (depthDiff < 0) {
-    // moving outward
-    fromParent = 0
-    fromCurrent = { inner: 0, outer: parentRadius }
-    fromChild = { inner: parentRadius, outer: r1 }
-  }
-
-  const currentRadii = useAnimatedRadii(targetCurrent, 1500, fromCurrent)
-
-  const prevRingRadii = useAnimatedRadii(targetPrevRing, 1500, fromPrevRing)
-
-  const childRadii = useAnimatedRadii(targetChild, 1500, fromChild)
-
-  const animatedParentRadius = useAnimatedNumber(targetParent, 1500, fromParent)
-
-  const parentOpacity = depthDiff > 0 ? parentFadeRaw : depthDiff < 0 ? 1 - parentFadeRaw : 1
-
-  if (tasks.length === 0) {
-    return <p>No tasks yet</p>
-  }
-
-  const parentName = parentPath.length ? getTaskByPath(tasks, parentPath)?.name : ''
-
-  return (
-    <svg className="pie" viewBox="-110 -110 220 220" width={220} height={220}>
-      <g
-        className="parent"
-        onClick={parentPath.length > 0 ? onUp : undefined}
-        style={{ opacity: parentOpacity }}
-      >
-        <circle cx={0} cy={0} r={animatedParentRadius} fill="#444" stroke="#000" />
-        {parentPath.length > 0 && (
-          <text x={0} y={0} textAnchor="middle" dominantBaseline="middle">
-            {parentName}
-          </text>
-        )}
-      </g>
-      {depthDiff > 0 && (
-        <g transform={`rotate(${prevRotationDeg})`} style={{ opacity: 1 - parentFadeRaw }}>
-          {prevCurrentTasks.map((task, i) => {
-            const { start, end } = prevAngles[i]
-            const pathD = describeRingArc(
-              0,
-              0,
-              prevRingRadii.inner,
-              prevRingRadii.outer,
-              start,
-              end,
-            )
-            const color = task.completed ? '#006400' : '#555'
-            return <path key={task.id} d={pathD} fill={color} stroke="#000" />
-          })}
-        </g>
-      )}
-      <g transform={`rotate(${rotationDeg})`} className="current">
-        {currentTasks.map((task, i) => {
-          const { start, end, mid } = angles[i]
-          const pathD = describeRingArc(
-            0,
-            0,
-            currentRadii.inner,
-            currentRadii.outer,
-            start,
-            end,
-          )
-          const color = task.completed ? '#006400' : '#555'
-          const textRadius = (currentRadii.inner + currentRadii.outer) / 2
-          const textPos = polarToCartesian(0, 0, textRadius, mid)
-          return (
-            <g key={task.id} onClick={() => onSelect([...parentPath, task.id])}>
-              <path
-                d={pathD}
-                fill={color}
-                stroke="#000"
-                strokeWidth={1}
-                className={task.id === selectedId ? 'selected' : undefined}
-              />
-              <text
-                x={textPos.x}
-                y={textPos.y}
-                textAnchor="middle"
-                transform={`rotate(${ -rotationDeg } ${textPos.x} ${textPos.y})`}
-              >
-                {task.name}
-              </text>
-            </g>
-          )
-        })}
-      </g>
-      {selectedTask && childTasks.length > 0 && (
-        <g transform={`rotate(${rotationDeg})`} className="children">
-          {childTasks.map((task, i) => {
-            const { start, end, mid } = childAngles[i]
-            const pathD = describeRingArc(0, 0, childRadii.inner, childRadii.outer, start, end)
-            const color = task.completed ? '#228b22' : '#777'
-            const textPos = polarToCartesian(0, 0, (childRadii.inner + childRadii.outer) / 2, mid)
-            return (
-              <g key={task.id} onClick={() => onSelect([...path, task.id])}>
-                <path d={pathD} fill={color} stroke="#000" />
-                <text
-                  x={textPos.x}
-                  y={textPos.y}
-                  textAnchor="middle"
-                  transform={`rotate(${ -rotationDeg } ${textPos.x} ${textPos.y})`}
-                >
-                  {task.name}
-                </text>
-              </g>
-            )
-          })}
-        </g>
-      )}
-    </svg>
-  )
-}
 
 let nextId = 1
