+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface Task {
+  id: number
+  name: string
+  description: string
+  effort: number
+  completed: boolean
+  subtasks: Task[]
+}
