+++ src/utils/useIsMobile.ts
@@ -1,0 +1,17 @@
+import { useEffect, useState } from 'react'
+
+export function useIsMobile() {
+  const query = '(max-width: 600px)'
+  const [isMobile, setIsMobile] = useState(() =>
+    typeof window === 'undefined' ? false : window.matchMedia(query).matches,
+  )
+
+  useEffect(() => {
+    const mq = window.matchMedia(query)
+    const handler = (e: MediaQueryListEvent) => setIsMobile(e.matches)
+    mq.addEventListener('change', handler)
+    return () => mq.removeEventListener('change', handler)
+  }, [])
+
+  return isMobile
+}
